+++ change-log.md
@@ -10,14 +10,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - move to dotnetcore - TBD.
 - updated selenium drivers
-- tests run a lot faster. (some numbers will be massively helpful.)
 - removed Winforms sendkeys, replaced with Selenium.Sendkeys.
 
 ### added
 
 ### deleted
 - removed phantomJs : See https://stackoverflow.com/questions/52442100/selenium-phantomjs-is-invalid-namespace
-
+- removed UploadFile. Simply use `EnterText` and `Click` to upload file instead, see `UploadTests.cs`.
 
 ## [1.1.0] - 2019-02-19